+++ packages/core-components/src/lib/unsorted/viz/core/ValueError.svelte
@@ -7,7 +7,7 @@
 </script>
 
 <span
-	class="z-50 group inline-flex items-center relative cursor-help cursor-helpfont-sans px-1 border border-red-200 py-[1px] bg-red-50 rounded"
+	class="group inline-flex items-center relative cursor-help cursor-helpfont-sans px-1 border border-red-200 py-[1px] bg-red-50 rounded"
 >
 	<span class="inline font-sans font-medium text-xs text-red-600">error</span>
 	<span
