+++ packages/core-components/src/lib/unsorted/viz/core/_Sparkline.svelte
@@ -1,17 +1,11 @@
 <script>
-	import { uiColours } from '@evidence-dev/component-utilities/colours';
 	import { onMount, onDestroy } from 'svelte';
-	import chroma from 'chroma-js';
 	import { init, connect } from 'echarts';
-	import {
-		formatValue,
-		getFormatObjectFromString
-	} from '@evidence-dev/component-utilities/formatting';
-	import getColumnSummary from '@evidence-dev/component-utilities/getColumnSummary';
 	import { browser } from '$app/environment';
 	import checkInputs from '@evidence-dev/component-utilities/checkInputs';
 	import ValueError from './ValueError.svelte';
 	import { strictBuild } from '@evidence-dev/component-utilities/chartContext';
+	import { getColumnFormats, getSparklineConfig, validateSize } from './sparkline.js';
 
 	export let config = {};
 	export let width = 50; // Default width for the sparkline, adjust as needed
@@ -27,12 +21,9 @@
 	export let dateCol = undefined;
 	export let valueCol = undefined;
 	export let valueFmt = undefined;
-	let valueFormat;
 	let value_format_object;
-	let columnSummary;
 
 	export let dateFmt = undefined;
-	let dateFormat;
 	let date_format_object;
 
 	export let type = 'line'; // line, area, or bar
@@ -61,239 +52,68 @@
 		}
 	}
 
-	$: try {
-		// ---------------------------------------------------------------------------------------
-		// Get column information
-		// ---------------------------------------------------------------------------------------
+	// Initialize the static SVG
+	onMount(() => {
+		if (!interactive) {
+			// Generate static SVG for non-interactive mode
+			const offscreenContainer = document.createElement('div');
+			offscreenContainer.style.width = width + 'px';
+			offscreenContainer.style.height = height + 'px';
+			const tempChart = init(offscreenContainer, 'evidence-light', {
+				renderer: 'svg',
+				height,
+				width
+			});
+			tempChart.setOption(config);
+			staticSVG = tempChart.renderToSVGString();
+			tempChart.dispose();
+		} else {
+			initializeChart();
+		}
+	});
 
+	// Cleanup
+	onDestroy(() => {
+		if (chartInstance) {
+			chartInstance.dispose();
+		}
+	});
+
+	$: try {
+		// Check that sparkline type is valid:
 		if (!['line', 'area', 'bar'].includes(type)) {
 			throw Error('type must be line, area, or bar');
 		}
 
-		if (height) {
-			// if height was user-supplied
-			height = Number(height);
-			if (isNaN(height)) {
-				// input must be a number
-				throw Error('height must be a number');
-			} else if (height <= 0) {
-				throw Error('height must be a positive number');
-			}
-		} else {
-			height = 15;
-		}
+		// Check that dimensions are valid:
+		({ height, width } = validateSize(height, width));
 
-		if (width) {
-			// if width was user-supplied
-			width = Number(width);
-			if (isNaN(width)) {
-				// input must be a number
-				throw Error('width must be a number');
-			} else if (width <= 0) {
-				throw Error('width must be a positive number');
-			}
-		} else {
-			width = 50;
-		}
-
+		// Check that inputs are valid:
 		checkInputs(data, [valueCol, dateCol]);
 
-		// Get column summary:
-		columnSummary = getColumnSummary(data);
-		// Get formats:
-		valueFormat = columnSummary[valueCol].format;
-		dateFormat = columnSummary[dateCol].format;
+		// Get column formats:
+		({ value_format_object, date_format_object } = getColumnFormats(
+			data,
+			valueCol,
+			dateCol,
+			valueFmt,
+			dateFmt
+		));
 
-		// Set column formats
-		value_format_object = valueFmt ? getFormatObjectFromString(valueFmt) : valueFormat;
-		date_format_object = dateFmt ? getFormatObjectFromString(dateFmt) : dateFormat;
-
+		// Prepare data for sparkline config:
 		const sparklineData = data.map((d) => [d[dateCol], d[valueCol]]);
 		sparklineData.sort((a, b) => a[0] - b[0]);
 
-		config = {
-			title: {
-				subtextStyle: {
-					width: '100%'
-				}
-			},
-			tooltip: {
-				trigger: 'axis',
-				position: function (point, params, dom, rect, size) {
-					// Calculate horizontal center and a fixed vertical offset
-					const horizontalCenter = size.viewSize[0] / 2 - size.contentSize[0] / 2;
-					const verticalOffset = -13.5; // Adjust this value to position the tooltip above the chart
-					return [horizontalCenter, verticalOffset];
-				},
-				formatter: function (params) {
-					// Assuming params[0] is your primary data point
-					const dataPoint = params[0];
-					// Customize these HTML blocks as needed
-					const valuePart = `<div style="text-align: center; background-color: transparent; border-radius: 1px; padding: 0px 2px; height: 12px;">${formatValue(
-						dataPoint.value[1],
-						value_format_object
-					)}</div>`;
-					const transparentGap = `<div style="height: ${height + 1.5}px;"></div>`; // Adjust height for the gap size
-					const datePart = `<div style="text-align: center; height: 1em; background-color: transparent; border-radius: 1px; padding: 0px 2px;">${formatValue(
-						dataPoint.axisValueLabel,
-						date_format_object
-					)}</div>`;
-
-					return valuePart + transparentGap + datePart;
-				},
-				backgroundColor: 'transparent', // Semi-transparent white background
-				borderWidth: 0,
-				borderColor: 'transparent',
-				extraCssText: 'box-shadow: none; padding-bottom: 0;', // Optional: Add some shadow for depth
-				padding: 0,
-				textStyle: {
-					fontSize: 10
-				}
-			},
-			legend: {
-				show: false,
-				type: 'scroll',
-				top: 0,
-				padding: [0, 0, 0, 0]
-			},
-			grid: {
-				// show: true,
-				left: 0,
-				right: 0,
-				bottom: 0,
-				top: 0,
-				containLabel: true
-			},
-			xAxis: {
-				type: 'time',
-				splitLine: {
-					show: false
-				},
-				axisLine: {
-					show: true,
-					lineStyle: {
-						width: 0.75,
-						color: uiColours.grey500
-					}
-				},
-				axisTick: {
-					show: false
-				},
-				axisLabel: {
-					show: false,
-					hideOverlap: true,
-					showMaxLabel: false,
-					formatter: false,
-					margin: 6
-				},
-				scale: true,
-				z: 2,
-				boundaryGap: '2%',
-				axisPointer: {
-					show: true,
-					snap: true,
-					type: 'line',
-					z: 0,
-					lineStyle: {
-						width: 0.5
-					},
-					handle: {
-						show: false
-					},
-					label: {
-						show: false
-					}
-				}
-			},
-			yAxis: [
-				{
-					type: 'value',
-					logBase: 10,
-					splitLine: {
-						show: false
-					},
-					axisLine: {
-						show: false,
-						onZero: false
-					},
-					axisTick: {
-						show: false
-					},
-					axisLabel: {
-						show: false,
-						hideOverlap: true,
-						margin: 4
-					},
-					name: '',
-					nameLocation: 'end',
-					nameTextStyle: {
-						align: 'left',
-						verticalAlign: 'top',
-						padding: [0, 5, 0, 0]
-					},
-					nameGap: 6,
-					scale: yScale,
-					boundaryGap: ['1%', '1%'],
-					z: 2
-				}
-			],
-			series: [
-				{
-					type: seriesType,
-					triggerLineEvent: true,
-					label: {
-						show: false,
-						position: 'top',
-						padding: 0,
-						fontSize: 9
-					},
-					labelLayout: {
-						hideOverlap: true
-					},
-					connectNulls: false,
-					emphasis: {
-						disabled: true
-					},
-					lineStyle: {
-						width: 1,
-						type: 'solid',
-						color: color ?? uiColours.grey400
-					},
-					areaStyle: {
-						color:
-							type === 'area'
-								? color
-									? chroma(color).brighten(1.5).hex()
-									: uiColours.grey300
-								: 'transparent'
-					},
-					itemStyle: {
-						color: color ?? uiColours.grey400
-					},
-					showSymbol: false,
-					symbol: 'circle',
-					symbolSize: 0,
-					step: false,
-					name: 'sparkline',
-					data: sparklineData,
-					yAxisIndex: 0
-				}
-			],
-			animation: false,
-			graphic: {
-				id: 'horiz-axis-title',
-				type: 'text',
-				style: {
-					text: '',
-					textAlign: 'right',
-					fill: 'hsla(212, 10%, 53%, 1)'
-				},
-				cursor: 'auto',
-				right: '3%',
-				top: null,
-				bottom: '2%'
-			}
-		};
+		config = getSparklineConfig(
+			sparklineData,
+			type,
+			seriesType,
+			color,
+			yScale,
+			value_format_object,
+			date_format_object,
+			height
+		);
 
 		if (!browser) {
 			// SSR-specific initialization
@@ -308,36 +128,9 @@
 			tempChart.dispose(); // Dispose instance after generating SVG
 		}
 
-		// Initialize the static SVG
-		onMount(() => {
-			if (!interactive) {
-				// Generate static SVG for non-interactive mode
-				const offscreenContainer = document.createElement('div');
-				offscreenContainer.style.width = width + 'px';
-				offscreenContainer.style.height = height + 'px';
-				const tempChart = init(offscreenContainer, 'evidence-light', {
-					renderer: 'svg',
-					height,
-					width
-				});
-				tempChart.setOption(config);
-				staticSVG = tempChart.renderToSVGString();
-				tempChart.dispose();
-			} else {
-				initializeChart();
-			}
-		});
-
 		if (chartContainer && interactive && !chartInstance) {
 			initializeChart();
 		}
-
-		// Cleanup
-		onDestroy(() => {
-			if (chartInstance) {
-				chartInstance.dispose();
-			}
-		});
 	} catch (e) {
 		error = e;
 		const setTextRed = '\x1b[31m%s\x1b[0m';
